First attempt at refactoring the transcriber module
- Read in f32 samples instead of s16 samples so that we don't need to do
  costly conversions
- Move the audio processing out of miniaudio's data callback
- Simplify the condition variables used in the SampleQueue

+++ include/transcriber/analyzer.h
@@ -8,11 +8,10 @@ using float_vec = std::vector<float>;
 class SpectrumAnalyzer {
 public:
   void init(int sample_rate);
-
-  float_vec process(int16_t* samples, size_t length);
+  float_vec process(float* samples, size_t length);
 
 private:
-  void fill_input_buffer(int16_t* data, size_t size);
+  void fill_input_buffer(float* data, size_t size);
   float_vec get_frequency_bins();
   float_vec map_bins_to_bars(float_vec bins);
